Add files via upload
E- Library is develop to carry out to keep the different type of
book where user can download the file of the book by paying certain amount of price.
The system has provided three different role i.e., ADMIN, AUTHOR, and USER.
This software's main goal is to easily access the book to the user as per their choice.
Here, In this system there is 2 type of user i.e. paid and free user, where paid user
can download the file as per the price of book and free user can only know the name
and price of book . In this system, the role of ADMIN is User Management and
Author role is to manage the Book price and upload the file of book. It also states
the methodology, tools used in the preparation of this report and the development of
the software system.